Parameterize the type of resource filters
Sometimes we would like our Resource filters to be `selects`, sometimes
we would rather have a text field as input. This allows parameterizing
them using the following syntax in the model:

```
def self.alchemy_resource_filters
  [
    {
      name: :start,
      type: :select,
      values: %w[starting_today future]
    },
    {
      name: :title,
      type: :text
    }
]
end

The `text` input is not implemented yet, but that's the idea. All other
keys in the configuration are simply passed on to the class that
implements the filter.

There is also `to_partial_path` implemented, so that each resource
filter class can define its own partial.

I could not just use `render resource_filters_for_select`, because Rails
has a global configuration option that would have required me to open a
new `admin` namespace for these partials otherwise. See
https://github.com/rails/rails/pull/5625 (This adds a config option to
turn that behavior off, but it's unfortunately global and incompatible
with engines).
